fix(rust): restore migrated workspace CI contracts

+++ server-rs/crates/cordy-cli/src/login_browser.rs
@@ -276,8 +276,10 @@ pub(crate) async fn wait_for_login_callback(
 pub(crate) fn constant_time_equal(left: &[u8], right: &[u8]) -> bool {
     let mut difference = left.len() ^ right.len();
     for index in 0..left.len().max(right.len()) {
-        difference |= left.get(index).copied().unwrap_or_default()
-            ^ right.get(index).copied().unwrap_or_default();
+        difference |= usize::from(
+            left.get(index).copied().unwrap_or_default()
+                ^ right.get(index).copied().unwrap_or_default(),
+        );
     }
     difference == 0
 }
